requirements.txt
torch
torchvision
torchaudio
pandas
numpy
jieba
matplotlib
seaborn
pyyaml
mlflow
scikit-learn 